#172 - fixes embedded links not being presented.
