--- ThunderCloud/EmbeddedLinksListItemCell.xib
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="13196" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="13173"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" rowHeight="138" id="KGk-i7-Jjw" customClass="EmbeddedLinksListItemCell" customModule="ThunderCloud" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="320" height="138"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
-                <rect key="frame" x="0.0" y="0.0" width="320" height="137.5"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" spacing="12" translatesAutoresizingMaskIntoConstraints="NO" id="MlI-T7-V9q">
-                        <rect key="frame" x="12" y="12" width="296" height="113.5"/>
-                        <subviews>
-                            <stackView opaque="NO" contentMode="scaleToFill" alignment="center" spacing="12" translatesAutoresizingMaskIntoConstraints="NO" id="1x5-Ca-jlu">
-                                <rect key="frame" x="0.0" y="0.0" width="296" height="51.5"/>
-                                <subviews>
-                                    <imageView userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="750" horizontalCompressionResistancePriority="1000" translatesAutoresizingMaskIntoConstraints="NO" id="B01-Gt-Ew8" customClass="ImageView" customModule="ThunderTable">
-                                        <rect key="frame" x="0.0" y="1" width="0.0" height="50"/>
-                                    </imageView>
-                                    <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" spacing="2" translatesAutoresizingMaskIntoConstraints="NO" id="toS-nj-qw5">
-                                        <rect key="frame" x="-12" y="4.5" width="0.0" height="43"/>
-                                        <subviews>
-                                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalCompressionResistancePriority="751" text="Title" textAlignment="natural" lineBreakMode="wordWrap" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="azA-O2-jDz">
-                                                <rect key="frame" x="0.0" y="0.0" width="0.0" height="20.5"/>
-                                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                                                <nil key="textColor"/>
-                                                <nil key="highlightedColor"/>
-                                            </label>
-                                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Subtitle" textAlignment="natural" lineBreakMode="wordWrap" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="oq9-h0-6p7">
-                                                <rect key="frame" x="0.0" y="22.5" width="0.0" height="20.5"/>
-                                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                                                <nil key="textColor"/>
-                                                <nil key="highlightedColor"/>
-                                            </label>
-                                        </subviews>
-                                    </stackView>
-                                </subviews>
-                            </stackView>
-                            <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" spacing="12" translatesAutoresizingMaskIntoConstraints="NO" id="8EE-v2-3kG">
-                                <rect key="frame" x="0.0" y="63.5" width="296" height="50"/>
-                            </stackView>
-                        </subviews>
-                    </stackView>
-                </subviews>
-                <constraints>
-                    <constraint firstItem="MlI-T7-V9q" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="12" id="OX6-nN-HnU"/>
-                    <constraint firstAttribute="bottom" secondItem="MlI-T7-V9q" secondAttribute="bottom" constant="12" id="feI-ah-yRD"/>
-                    <constraint firstAttribute="trailing" secondItem="MlI-T7-V9q" secondAttribute="trailing" constant="12" id="nOU-gd-r6W"/>
-                    <constraint firstItem="MlI-T7-V9q" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="12" id="tk0-bN-NP7"/>
-                </constraints>
-            </tableViewCellContentView>
-            <connections>
-                <outlet property="cellDetailLabel" destination="oq9-h0-6p7" id="yWH-t2-8Z2"/>
-                <outlet property="cellImageView" destination="B01-Gt-Ew8" id="iEW-ke-JiA"/>
-                <outlet property="cellTextLabel" destination="azA-O2-jDz" id="zzT-jy-S0Y"/>
-                <outlet property="contentStackView" destination="1x5-Ca-jlu" id="MFy-Ol-dDH"/>
-                <outlet property="embeddedLinksStackView" destination="8EE-v2-3kG" id="WL3-sM-Fje"/>
-                <outlet property="mainStackView" destination="MlI-T7-V9q" id="OVl-bF-tBb"/>
-            </connections>
-            <point key="canvasLocation" x="-182" y="51"/>
-        </tableViewCell>
-    </objects>
-</document>
