+++ ThunderCloud/LinkCollectionCell.swift
@@ -14,7 +14,7 @@ import StoreKit
 class LinkCollectionCell: CollectionCell {
 	
 	/// The array of links to be shown in the collection view
-	var links: [LinkCollectionItem]? {
+	var linkItems: [LinkCollectionItem]? {
 		didSet {
 			reload()
 		}
@@ -51,7 +51,7 @@ extension LinkCollectionCell {
 	override func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
 		
 		let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "Cell", for: indexPath)
-		guard let links = links, let linkCell = cell as? LinkScrollerCollectionViewCell else { return cell }
+		guard let links = linkItems, let linkCell = cell as? LinkScrollerCollectionViewCell else { return cell }
 		
 		let link = links[indexPath.row]
 		linkCell.imageView.image = link.image
@@ -79,7 +79,7 @@ extension LinkCollectionCell {
 	
 	func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
 		
-		guard let links = links, let link = links[indexPath.item].link else { return }
+		guard let links = linkItems, let link = links[indexPath.item].link else { return }
 		parentViewController?.navigationController?.push(link: link)
 	}
 }
